Fixes #7: msgpack deserialization broken
The msgpack classes weren't being loaded, as they weren't
required by any of the namespaces that were being loaded.

Previously, it was thought that adding kbrowse.msgpack as an :aot
in the project configuration would resolve, but this was not the
case, and so has been removed.

This also adds an integration test that fails w/o the fix.