Fixes #39: Topics stop refreshing
#### Background
During initialization, kbrowse starts a separate thread that intermittenly
refreshes the topics lists from the kafka brokers.

If an exception was thrown during a refresh, the thread died,
and topics would stop refreshing.

This could happen if a cluster were down for long enough to trigger a
timeout exception, for example.

#### Fix
This adds exception handling around the refresh call.

If we catch an exception, we log it, but continue on.

#### Test Plan
1. Start Kafka/Zookeeper:
```
./run-zookeeper-and-kafka
```

2. Start kbrowse:
```
KAFKA_TOPICS_CACHE_SLEEP_SECONDS=5 ./lein run server
```

3. Ctrl-C Kafka/Zookeeper

4. Wait for kbrowse timeout to log

5. Restart Kafka/Zookeeper

6. Create new topic:
```
./kafka/bin/kafka-topics.sh --zookeeper localhost:2181 --create --topic kbrowse-2 --partitions 10 --replication-factor 1
```

7. Verify that the new topic shows up in running kbrowse:
```
http://localhost:4000
```